Data converted to JSON, Comment, Edits ,Animation Work
the data has now been converted into a JSON file and is ready to be used from it. i removed some commented out code as well as unnecessary comments. display for tooltip is now none so it will not affect pre-existing html elements in the LabCentral Micro-Site and Main Site. The tooltip will only exist when a job node is hovered over